feat: 미니 QA 진행 후 코드 수정
