+++ src/server/requestTypes/member-stat.ts
@@ -30,11 +30,57 @@ export interface RegisterMemberStatRequest {
 }
 
 export interface GetFilteredMemberListRequest {
-  [key: string]: (number | string)[];
+  birthYear: number[];
+  admissionYear: string[];
+  majorName: string[];
+  acceptance: string[];
+  wakeUpTime: number[];
+  sleepingTime: number[];
+  turnOffTime: number[];
+  smoking: string[];
+  sleepingHabit: string[];
+  airConditioningIntensity: number[];
+  heatingIntensity: number[];
+  lifePattern: string[];
+  intimacy: number[];
+  canShare: string[];
+  isPlayGame: string[];
+  isPhoneCall: string[];
+  studying: string[];
+  intake: string[];
+  cleanSensitivity: number[];
+  noiseSensitivity: number[];
+  cleaningFrequency: string[];
+  drinkingFrequency: string[];
+  personality: string[];
+  mbti: string[];
 }
 
 export interface GetFilteredMemberListCountRequest {
-  [key: string]: (number | string)[];
+  birthYear: number[];
+  admissionYear: string[];
+  majorName: string[];
+  acceptance: string[];
+  wakeUpTime: number[];
+  sleepingTime: number[];
+  turnOffTime: number[];
+  smoking: string[];
+  sleepingHabit: string[];
+  airConditioningIntensity: number[];
+  heatingIntensity: number[];
+  lifePattern: string[];
+  intimacy: number[];
+  canShare: string[];
+  isPlayGame: string[];
+  isPhoneCall: string[];
+  studying: string[];
+  intake: string[];
+  cleanSensitivity: number[];
+  noiseSensitivity: number[];
+  cleaningFrequency: string[];
+  drinkingFrequency: string[];
+  personality: string[];
+  mbti: string[];
 }
 
 export interface UpdateMemberStatRequest {
